Add a Java 9+ MultiMC instance
Thanks to the efforts of @LightWayUp on LightWayUp/LegacyLauncher, we can use LaunchWrapper on Java 9+! RetroWrapper is apparently compatible with ASM 5.2, so I've taken the liberty of upgrading the versions used in the Java 9+ instance. After that, some of the internal hacks it does breaks things, but that's something I'll solve another day.